% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{addmodulescore_multiple_pathway}
\alias{addmodulescore_multiple_pathway}
\title{addmodulescore_multiple_pathway}
\usage{
addmodulescore_multiple_pathway(d3, gmt_list, name)
}
\arguments{
\item{d3}{seurat_object}

\item{gmt_list}{gmt list}

\item{name}{title of the data}
}
\value{
signature score of pathways by gene expression for each clone in the sample (multiple pathways)
}
\description{
measure signature score for each clone in the sample with multiple pathways
}
\examples{
gmt_list <- list()
gmt <- getGmt("cancersea_gmt")
gmt_list$cancersea<- gmt
gmt <- getGmt("core_gmt")
gmt_list$core_fitness <- gmt
gmt <- getGmt("lung_cm_gmt")
gmt_list$cancermine <- gmt
gmt <- getGmt("lung_gmt")
gmt_list$specific_fitness <- gmt
gmt <- getGmt("lung_gmt")
gmt_list$ctg <- gmt
addmodulescore_multiple_pathway(seurat_object, gmt_list, "lung")

}
