% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{avg_funct}
\alias{avg_funct}
\title{avg_funct}
\usage{
avg_funct(a1)
}
\arguments{
\item{a1}{path(row)-feature(column) table}
}
\value{
mean feature score for each path
}
\description{
averaging multiple feature scores
}
\examples{
a<-total_table[,1:feature_size,drop=F]
a1<-a[,cluster_member, drop = F]
a2 <- avg_funct(a1)

}
