% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution_path_analysis.R
\name{clone_cluster_weighted_lr_abundance}
\alias{clone_cluster_weighted_lr_abundance}
\title{clone_cluster_weighted_lr_abundance}
\usage{
clone_cluster_weighted_lr_abundance(
  meta_info,
  name,
  pat_info,
  sample_info,
  cluster_info,
  clone_path_file_prefix,
  feature_file_prefix,
  abundance_file_prefix,
  output_file_prefix,
  cell_cnt_thr = 10
)
}
\arguments{
\item{meta_info}{seurat_obj@meta.data}

\item{name}{title for your data}

\item{pat_info}{variable name for patient in the "meta_info"}

\item{sample_info}{variable name for sample in the "meta_info"}

\item{cluster_info}{expression based cluster variable name for sample in the "meta_info" (ex: seurat_clusters)}

\item{clone_path_file_prefix}{prefix of clone path file}

\item{feature_file_prefix}{prefix of feature file}

\item{abundance_file_prefix}{prefix of abundance file}

\item{output_file_prefix}{prefix of outputfile}

\item{cell_cnt_thr}{threshold for excluding clones less than the given cell count}
}
\value{
Correlation analysis between each evolution path and mutational features
}
\description{
Correlation analysis between clonal abundance and score from LR in each clone in each evolution path
}
\examples{
#meta_info <- seurat_obj@meta.data
meta_info <- read.table("data/meta_info_lung", sep = "\t", check.names = F)
name <- "lung"
pat_info <- "pat_collapse"
sample_info <- "sample_name"
cluster_info <- "seurat_clusters"
clone_path_file_prefix <- "data/robustclone/"
feature_file_prefix <- "data/cellchat/"
abundance_file_prefix <- "data/"
output_file_prefix <- "./"
clone_cluster_weighted_lr_abundance(meta_info, name, pat_info, sample_info, cluster_info, clone_path_file_prefix, feature_file_prefix, abundance_file_prefix, output_file_prefix)

}
