% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution_path_analysis.R
\name{clone_path_list}
\alias{clone_path_list}
\title{clone_path_list}
\usage{
clone_path_list(sample_clone_path)
}
\arguments{
\item{sample_clone_path}{total evolution path (igraph) object}
}
\value{
each evolution path in a list object
}
\description{
split each evolution path from total evolution path (igraph) object
}
\examples{
clone_path <- read.table("data/robustclone/TH155_clone_path")
clone_path <- as.matrix(clone_path)
clone_path[1,] <- as.character(clone_path[1,])
clone_path[2,] <- as.character(clone_path[2,])
clone_path_graph <-graph_from_edgelist(clone_path)
sample_clone_path <- subgraph(clone_path_graph, sample_clone_list)
#sample_clone_list: clone list in that sample
total_path <- clone_path_list(sample_clone_path)

}
