% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{clone_shannon_process}
\alias{clone_shannon_process}
\title{clone_shannon_process}
\usage{
clone_shannon_process(tmp_d, filter_cnt, sample)
}
\arguments{
\item{tmp_d}{seurat object with "seurat_clusters" and "cancer_clone"}

\item{filter_cnt}{threshold of cell count for each clone}

\item{sample}{sample name}
}
\value{
entropy vector
}
\description{
Evaluating entropy for each clone in a given sample
}
\examples{
d3 <- seurat_obj
d3 <- FindClusters(d3, resolution = resolution, verbose = FALSE)
tmp_d <- subset(d3, subset = cancer_clone != "NA")
tmp_res1 <- shannon_process(tmp_d, 0, "sample1")

}
