% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{clustering_plot}
\alias{clustering_plot}
\title{clustering_plot}
\usage{
clustering_plot(
  g6,
  clustering_res,
  name,
  layout = "layout_with_fr",
  format = "png",
  width = 1600,
  height = 900,
  resolution = 60
)
}
\arguments{
\item{g6}{clustered feature igraph object}

\item{clustering_res}{module_info from clustering function}

\item{name}{tilte of the data}

\item{layout}{layout for plot
\itemize{
\item "layout_with_fr"
\item "layout_with_graphopt"
}}

\item{format}{png or pdf}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{resolution}{resolution of png plot}
}
\value{
image
}
\description{
plot clustered feature graph
}
\examples{
[1: plotting connected componenets]
new_result <- feature_clustering_connected_comp(g3, total_table_pre, feature_size_pre) 
total_table_new<-new_result[[1]]
module_info_new <- new_result[[2]]
feature_size_new <- length(module_info_new)
g6<-new_result[[3]]
clustering_res <- new_result[[4]]
clustering_plot(g6,clustering_res,"myplot.png")

[2: plotting louvain clustering]
new_result1 <- feature_clustering_louvain(g6, total_table_pre, feature_size_pre, resolution = 1.1)
total_table<-new_result1[[1]]
module_info <- new_result1[[2]]
feature_size <- length(module_info)
g3<-new_result1[[3]]
g4<-new_result1[[4]]
clustering_plot(g3,g4,"myplot.png")
}
