% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{common_mutation_from_path}
\alias{common_mutation_from_path}
\title{common_mutation_from_path}
\usage{
common_mutation_from_path(path_list_info, feature_gn, root_exclude)
}
\arguments{
\item{path_list_info}{list object with sample, path, sample_clone_mut_list for each sample}

\item{feature_gn}{gene list for a given feature
\itemize{
\item "all": calling all the mutation from a given clone
\item gene_vector: calling only the mutation within a given gene list
}}

\item{root_exclue}{whether user wants to exclude mutation profile of root state or not
\itemize{
\item True
\item False
}}
}
\value{
return overlapping genelist
}
\description{
extracting overlapping mutational profile in multiple evolution path
}
\examples{
path_list_info<-list()
a<- c("LT_S21__29_23_27_25", "LT_S53__3_1_2")

for (j in c(1:length(a))){
	path <- a[j]
	sample<-strsplit(a[1], "__")[[1]][1]
	file_name <- paste(sample, "_clone_mut_list",sep = "")
	load(file_name)
	path_list_info[[j]] <- list(sample, sample_clone_mut_list)
	names(path_list_info)[j] <- path
} 

#names(path_list_info): path
#path_list_info[[1]][[1]]: sample
#path_list_info[[1]][[2]]: sample_clone_mut_list

genes <- common_mutation_from_path(path_list_info, "all", F)
}
