% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{feature_clustering_connected_comp}
\alias{feature_clustering_connected_comp}
\title{feature_clustering_connected_comp}
\usage{
feature_clustering_connected_comp(
  g3,
  total_table,
  feature_size,
  z_index = T,
  z_thr = 2.576
)
}
\arguments{
\item{g3}{result of feature_graph}

\item{total_table}{result of generate_table_for_analysis}

\item{feature_size}{number of features}

\item{z_index}{conduct z normalization on distances for the thresholding\
\itemize{
\item TRUE: use z-score
\item FALSE: use raw distance score
}}

\item{z_thr}{threshold of z-score}
}
\value{
list(new_a, module_info, g6, g4)
\itemize{
\item new_a: new table of mean feature score in each clusters
\item module_info: list object contains cluster member information for each feature
\item g6: igraph object after thresholding
\item g4: g4<-components(g6)
}
}
\description{
clustering features by connected component after thresholding
}
\examples{
new_result <- feature_clustering_connected_comp(g3, total_table_pre, feature_size_pre)
}
