% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{feature_clustering_louvain}
\alias{feature_clustering_louvain}
\title{feautre_clustering_louvain}
\usage{
feature_clustering_louvain(g3, total_table, feature_size, resolution = 1.1)
}
\arguments{
\item{g3}{result of feature_graph}

\item{total_table}{result of generate_table_for_analysis}

\item{feature_size}{number of the features}

\item{resolution}{resolution for louvain clustering}
}
\value{
list(new_a, module_info, g3, g4)
\itemize{
\item new_a: new table of mean feature score in each clusters
\item module_info: list object contains cluster member information for each feature
\item g3: initial input igraph object
\item g4: g4<-components(g6) \link{defining connected components by igraph}
}
}
\description{
Louvain clustering for feature graph
}
\examples{

}
