% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{feature_graph}
\alias{feature_graph}
\title{feature_graph}
\usage{
feature_graph(total_table, feature_size)
}
\arguments{
\item{total_table}{result of generate_table_for_analysis}

\item{feature_size}{number of features}
}
\value{
igraph object of features
}
\description{
make feature table into igraph object
}
\examples{
result <- generate_table_for_analysis(name, meta_info, pat_info, sample_info, meta_info_list,file_prefix,file_suffix)
total_table_pre <- result[[1]]
feature_size_pre <- result[[2]]
g3 <-feature_graph(total_table_pre,  feature_size_pre)
 
}
