% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_analysis.R
\name{generate_table_for_analysis}
\alias{generate_table_for_analysis}
\title{generate_table_for_analysis}
\usage{
generate_table_for_analysis(
  name,
  meta_info,
  pat_info,
  sample_info,
  meta_info_list,
  file_prefix,
  file_suffix,
  mut_trans_index = FALSE
)
}
\arguments{
\item{name}{title for the data}

\item{meta_info}{meta information}

\item{pat_info}{variable name for patient}

\item{sample_info}{variable name for sample}

\item{meta_info_list}{list of meta information to analyze}

\item{file_prefix}{prefix of the file}

\item{file_suffix}{suffix of the file}

\item{mut_trans_index}{whether feature is mut-trans coassociation}
}
\value{
total_table for statistical analysis
}
\description{
making a table for statistic analysis
}
\examples{
#meta_info <- seurat_obj@meta.data
meta_info <- read.table("data/meta_info_lung", sep = "\t", check.names = F)
name <- "lung"
pat_info <- "pat_collapsed"
sample_info <- "sample_name"
meta_info_list <- c("primary_or_metastaic","biopsy_site","smokingHx","histolgy","best_rxn_status")
file_prefix <- "./"
file_suffix <- "_clone_abundance_mut_scc_sample"
result_suffix <- "abundance_mut_scc_analysis"
result <- generate_table_for_analysis(name, meta_info, pat_info, sample_info, meta_info_list,file_prefix, file_suffix)
total_table <- result[[1]]
feature_size <- result[[2]]
}
