% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{lr_gmt}
\alias{lr_gmt}
\title{lr_gmt}
\usage{
lr_gmt(source, target, name)
}
\arguments{
\item{source}{list object of (ligand gene from cancer)}

\item{target}{list object of (receptor gene from cancer)}

\item{name}{title of the data}
}
\value{
list variable composed of geneset for each source/target of counter_celltype
}
\description{
make cellchat result into list format
}
\examples{
name <- "lung"

load(paste("../cellchat/",name, "_target_gn", sep = ""))
target<- res
load(paste("../cellchat/",name, "_source_gn", sep = ""))
source<- res
res <- lr_gmt(source, target, name)
save(res, file = paste(name, "_cellchat_gmt", sep = "")) 

}
