% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{lr_module}
\alias{lr_module}
\title{lr_module}
\usage{
lr_module(res, d, total_gn)
}
\arguments{
\item{res}{object from lr_gmt}

\item{d}{seurat object}

\item{total_gn}{total gene list in the seurat object}
}
\value{
signature score of LR by gene expression for each cancer cell in the sample
}
\description{
addmodulescore of each source/target gene in each cancer cell
}
\examples{
name <- "lung"
print(name)
load(paste(name, "_cellchat_gmt", sep = ""))
#cellchat gmt object name: res (from lr_gmt)
load(paste(name, "_cancer_seurat_mut", sep = ""))
#seurat object name: d3
total_gn <- rownames(d3)
res2 <- lr_module(res, d3,total_gn)
write.table(res2, paste(name,"_cell_addmodule", sep = ""), quote = F, sep = "\t")

}
