% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{mut_gmt}
\alias{mut_gmt}
\title{mut_gmt}
\usage{
mut_gmt(sample_list, gmt_list, clone_info_prefix, clone_info_suffix)
}
\arguments{
\item{sample_list}{list of samples}

\item{gmt_list}{multiple pathways in GSEABase gmt object}

\item{clone_info_prefix}{prefix of clone_info}

\item{clone_info_suffix}{suffix of clone_info}
}
\value{
jaccard coefficient between mutational profile and multiple pathways for each clone in the sample
}
\description{
measuring jaccard coefficient between mutational profile and multiple pathways for each clone in the sample
}
\examples{
gmt_list <- list()
gmt <- getGmt("cancersea_gmt")
gmt_list$cancersea<- gmt
gmt <- getGmt("core_gmt")
gmt_list$core_fitness <- gmt
gmt <- getGmt("lung_cm_gmt")
gmt_list$cancermine <- gmt
gmt <- getGmt("lung_gmt")
gmt_list$specific_fitness <- gmt
gmt <- getGmt("lung_gmt")
gmt_list$ctg <- gmt
mut_gmt(sample_list, gmt_list, "data/clone_mutation_info/", "_clone_mut_list")
}
