% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{mut_gmt_module}
\alias{mut_gmt_module}
\title{mut_gmt_module}
\usage{
mut_gmt_module(
  exp,
  meta_info,
  sample_info,
  gmt_list,
  clone_info_prefix,
  clone_info_suffix,
  name
)
}
\arguments{
\item{exp}{expression matrix}

\item{meta_info}{seurat_obj@meta.data}

\item{sample_info}{variable name for sample in the "meta_info"}

\item{gmt_list}{gmt list}

\item{clone_info_prefix}{prefix of clone_info}

\item{clone_info_suffix}{suffix of clone_info}

\item{name}{title of the data}
}
\value{
mean expression of pathways by (mutated) gene expression for each clone in the sample (multiple pathways)
}
\description{
mean expression of mutated gene of a given pathway in each clone
}
\examples{
gmt_list <- list()
gmt <- getGmt("cancersea_gmt")
gmt_list$cancersea<- gmt
gmt <- getGmt("core_gmt")
gmt_list$core_fitness <- gmt
gmt <- getGmt("lung_cm_gmt")
gmt_list$cancermine <- gmt
load("your_seurat_object")
#object name: d3
sample_info <- "sample_name"
exp <- GetAssayData(d3)
meta_info <- d3@meta.data
name <- "lung"
clone_info_prefix <- "data/clone_mutation_info/"
clone_info_suffix <- "_clone_mut_list"
mut_gmt_module(exp, meta_info, sample_info, gmt_list, clone_info_prefix,clone_info_suffix, name)
}
