% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{path_gmt}
\alias{path_gmt}
\title{path_gmt}
\usage{
path_gmt(gmt, gn)
}
\arguments{
\item{gmt}{gmt object from GSEABase package}

\item{gn}{input genelist}
}
\value{
jaccard coefficient between input gene and geneset from a given gmt (pathway)
}
\description{
Evaulate jaccard coefficient between input gene and geneset from a given gmt (pathway)
}
\examples{
tmp_gmt_res <- path_gmt(gmt, gn)
}
