% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{path_gmt_lr}
\alias{path_gmt_lr}
\title{path_gmt_lr}
\usage{
path_gmt_lr(res, gn, pre_name)
}
\arguments{
\item{res}{list object contains receptor (pre_name: target) or ligand (pre_name: sourcr) list of each cancer clone between each celltype}

\item{gn}{input gene}

\item{pre_name}{"target" (receptor gene from cancer) or "source" (ligand gene from cancer)}
}
\value{
jaccard coefficient for each interaction
}
\description{
Evaulate jaccard coefficient between ligand-receptor gene and a given geneset
}
\examples{
load("data/cellchat/lung_target_gn")
#object name: res
tmp_res <- path_gmt_lr(res, gn, "target")

load("data/cellchat/lung_source_gn")
#object name: res
tmp_res <- path_gmt_lr(res, gn, "source")

}
