% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_generation.R
\name{path_module}
\alias{path_module}
\title{path_module}
\usage{
path_module(gmt, d, total_gn)
}
\arguments{
\item{gmt}{gmt object from GSEABase package}

\item{d}{seurat object}

\item{total_gn}{total gene list in the data}
}
\value{
signature score for each clone by gene expression in the sample
}
\description{
measure signature score for each clone by gene expression in the sample with a given pathway
}
\examples{
gmt <- gmt_list[[z]]
tmp_res <- path_module(gmt, d3, total_gn)
}
