% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{random_entropy}
\alias{random_entropy}
\title{random_entropy}
\usage{
random_entropy(cluster_vec)
}
\arguments{
\item{cluster_vec}{cluster(clone) vector for a given sample}
}
\value{
entropy
}
\description{
shannon entropy for random expectation
}
\examples{
ent <- random_entropy(as.numeric(obj$seurat_clusters))
}
