% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_analysis.R
\name{sample_table_making_mut_trans}
\alias{sample_table_making_mut_trans}
\title{sample_table_making_mut_trans}
\usage{
sample_table_making_mut_trans(
  tmp_file_name,
  sample_name,
  sample_position,
  meta_table
)
}
\arguments{
\item{tmp_file_name}{file name of evolution analysis}

\item{sample_name}{sample name}

\item{sample_position}{index of given sample from total single-cell data}

\item{meta_table}{table for meta information}
}
\value{
table for the statistic analysis in evolution analysis (mut_trans analysis)
}
\description{
making table for the statistic analysis in evolution analysis (mut_trans analysis)
}
\examples{
sample_list <- as.character(meta_info[,sample_info])
sample_list_uniq <- unique(sample_list)
meta_table<- c()
for (j in c(1:length(meta_info_list))){
 meta_info_tmp <- meta_info_list[j]
 meta_list <- as.character(meta_info[,meta_info_tmp])
 meta_table <- rbind(meta_table, meta_list)
 rownames(meta_table)[j] <- meta_info_tmp
 }
 sample_name <- sample_list_uniq[1]
 sample_position <- which(sample_list == sample_name)
 tmp_file_name <- paste(file_prefix, sample_name,file_suffix, sep = "")
 a2<-sample_table_making_mut_trans(tmp_file_name, sample_name, sample_position, meta_table)

}
