% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{select_path_by_thr}
\alias{select_path_by_thr}
\title{select_path_by_thr}
\usage{
select_path_by_thr(total_table, sample, feature, thr, sign)
}
\arguments{
\item{total_table}{output of evolution path analysis}

\item{sample}{sample of interest}

\item{feature}{feature of interest}

\item{thr}{threshold for feature score}

\item{sign}{selecting higher or lower paths than the threshold
\itemize{
\item high (>thr)
\item low (<= thr)
}}
}
\value{
selected evolution path vector
}
\description{
select evolution paths by a given threshold of a given feature
}
\examples{
path_list <- select_path_by_thr(total_table,  "LT_S21", "Angiogenesis",1, "high")
}
