% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{shannon_entropy}
\usage{
shannon_entropy(x, batch_vector, N_batches, cluster)
}
\arguments{
\item{x}{given clone or cluster}

\item{batch_vector}{cluster vector or clone vector (batch)}

\item{N_batches}{number of cluster or clone}

\item{cluster}{clone vector or cluster vector}
}
\value{
entropy
}
\description{
normalized shannon entropy
}
\examples{
ent <- shannon_entropy(clone, cluster_vec, cluster_size, clone_vec)
}
