% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_analysis.R
\name{statistic_test}
\alias{statistic_test}
\title{statistic_test}
\usage{
statistic_test(
  total_table,
  feature_size,
  meta_info_input,
  method,
  name,
  result_suffix,
  plot_index,
  format = "png",
  width = 640,
  height = 1200,
  resolution = 80
)
}
\arguments{
\item{total_table}{result of generate_table_for_analysis}

\item{feature_size}{number of the features}

\item{meta_info_input}{name of the meta information of interest}

\item{method}{wilcoxon-rank sum test or Kruskal-wallis test}

\item{name}{tile of the data}

\item{result_suffix}{suffix of the result file}

\item{plot_index}{plot (TRUE or FALSE)}

\item{format}{png or pdf}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{resolution}{resolution of png plot}
}
\value{
statistical result and image (if applied)
}
\description{
statistical test with a given meta information in the dataset (+ plotting)
}
\examples{
result_suffix <- "evolution_mut_scc_pathwise_analysis"
result <- generate_table_for_analysis(name, meta_info, pat_info, sample_info, meta_info_list,file_prefix, file_suffix)
total_table <- result[[1]]
feature_size <- result[[2]]
statistic_res<-statistic_test(total_table, feature_size, "biopsy_site", "kruskal", "lung", result_suffix, T)

statistic_res<-statistic_test(total_table, feature_size, "histolgy", "wilcox", "lung", result_suffix, T)

}
